third_party/blink/web_tests/external/wpt/js-self-profiling/max-buffer-size.https.html

<!DOCTYPE html>
<html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>

  <script src="resources/profile-utils.js"></script>
</head>
<body>
  <script>
    promise_test(async t => {
      assert_throws_js(TypeError, () => {
        new Profiler({ sampleInterval: 10 });
      });
    }, 'max buffer size must be defined');

    promise_test(async t => {
      const profiler = new Profiler({
        sampleInterval: 10,
        maxBufferSize: 2,
      });

      // Force 3 samples with a max buffer size of 2.
      for (let i = 0; i < 3; i++) {
        ProfileUtils.forceSample();
      }

      const trace = await profiler.stop();
      assert_equals(trace.samples.length, 2);
    }, 'max buffer size is not exceeded');

    promise_test(async t => {
      const pf = [];
      pf[0] = new Profiler({ sampleInterval: 10,  maxBufferSize: Number.MAX_SAFE_INTEGER });
      pf[1] = new Profiler({ sampleInterval: 10,  maxBufferSize: 1 });

      const watcher = new EventWatcher(t, pf[1], ['samplebufferfull']);
      pf[0].addEventListener("samplebufferfull", ()=>{
        assert_unreached("samplebufferfull invoked on wrong profiler");
      });
      pf[1].addEventListener("samplebufferfull", async ()=>{
        pf[0].stop();
        pf[1].stop();
        assertTrue(true);
      });

      for (let i = 0; i < 2; i++) {
        ProfileUtils.forceSample();
      }

      return watcher.wait_for('samplebufferfull');
    }, 'executed samplebufferfull function');
  </script>
</body>
</html>
